+++ eras/conway/test-suite/src/Test/Cardano/Ledger/Conway/Serialisation/Roundtrip.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
+{-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
@@ -6,21 +7,32 @@
 
 module Test.Cardano.Ledger.Conway.Serialisation.Roundtrip (allprops) where
 
+import Cardano.Ledger.Alonzo.TxWits (Redeemers)
+import Cardano.Ledger.Binary.Version (natVersion)
 import Cardano.Ledger.Conway.Genesis (ConwayGenesis (..))
 import Cardano.Ledger.Core (Era (..))
 import Data.Data (Proxy (..), typeRep)
-import Test.Cardano.Ledger.Binary.RoundTrip (roundTripCborExpectation)
+import Test.Cardano.Ledger.Binary.RoundTrip (
+  embedTripAnnExpectation,
+  roundTripAnnRangeExpectation,
+  roundTripCborExpectation,
+ )
 import Test.Cardano.Ledger.Conway.Serialisation.Generators ()
+import Test.Hspec (shouldBe)
 import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.QuickCheck (testProperty)
 
 allprops ::
   forall e.
-  ( Era e
-  ) =>
+  (Era e) =>
   TestTree
 allprops =
   testGroup
-    (show $ typeRep (Proxy @e))
-    [ testProperty "ConwayGenesis" $ roundTripCborExpectation @(ConwayGenesis (EraCrypto e))
+    ("Conway CBOR roundtrip " <> show (typeRep (Proxy @e)))
+    [ testProperty "ConwayGenesis" $
+        roundTripCborExpectation @(ConwayGenesis (EraCrypto e))
+    , testProperty "Redeemers" $
+        roundTripAnnRangeExpectation @(Redeemers e) (natVersion @9) maxBound
+    , testProperty "Redeemers - Conway can decode a Babbage encoding" $
+        embedTripAnnExpectation @(Redeemers e) (natVersion @8) (natVersion @9) shouldBe
     ]